archive/products/omniapixels/OmniaPixelsApp-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.34.0
minio==7.2.0
bcrypt==4.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
coverage==7.3.2
ruff==0.1.6
email-validator==2.1.0.post1
Pillow==10.4.0
# Sprint-2: Enhanced image processing
numpy==1.24.3
opencv-python==4.8.1.78
